requirement.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
backcall==0.1.0
bleach==3.1.0
blis==0.4.1
catalogue==0.0.8
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
en-core-web-sm==2.2.0
entrypoints==0.3
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
future==0.17.1
fuzzywuzzy==0.17.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.8
imageio==2.6.1
importlib-metadata==0.23
imutils==0.5.3
incremental==16.10.1
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.0.2
jupyter-client==5.3.3
jupyter-core==4.5.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
murmurhash==1.0.2
nbconvert==5.6.0
nbformat==4.4.0
ner-d==0.2
netifaces==0.10.4
networkx==2.4
notebook==6.0.1
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.1.26
PAM==0.4.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2==2.8.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
Pygments==2.4.2
pygobject==3.26.1
PyJWT==1.7.1
pyOpenSSL==17.5.0
pyparsing==2.4.4
pyrsistent==0.15.4
pyserial==3.4
pytesseract==0.3.0
python-apt==1.6.4
python-dateutil==2.8.1
python-debian==0.1.32
python-Levenshtein==0.12.0
python-twitter==3.5
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.12
pyzmq==18.1.0
requests==2.22.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
scikit-image==0.16.2
scipy==1.3.1
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==16.0.0
six==1.13.0
spacy==2.2.3
SQLAlchemy==1.3.10
srsly==0.2.0
ssh-import-id==5.7
systemd-python==234
terminado==0.8.2
tesseract==0.1.3
tesserocr==2.4.1
testpath==0.4.2
thinc==7.3.1
tornado==6.0.3
tqdm==4.38.0
traitlets==4.3.2
twilio==6.33.1
Twisted==17.9.0
ufw==0.36
urllib3==1.25.7
virtualenv==15.1.0
wasabi==0.4.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
xx-ent-wiki-sm==2.2.0
zipp==0.6.0
zope.interface==4.3.2
